Clean up deps in pkg/inngest
This allows for clean imports.
